+++ sdk/src/testIntegration/java/com/hedera/hashgraph/sdk/test/integration/AccountRecordsIntegrationTest.java
@@ -20,7 +20,7 @@ class AccountRecordsIntegrationTest {
             var key = PrivateKey.generateED25519();
 
             var response = new AccountCreateTransaction()
-                    .setKey(key)
+                    .setKeyWithoutAlias(key)
                     .setInitialBalance(new Hbar(1))
                     .execute(testEnv.client);