+++ sdk/src/testIntegration/java/com/hedera/hashgraph/sdk/test/integration/TokenFreezeIntegrationTest.java
@@ -26,7 +26,7 @@ class TokenFreezeIntegrationTest {
             var key = PrivateKey.generateED25519();
 
             var response = new AccountCreateTransaction()
-                    .setKey(key)
+                    .setKeyWithoutAlias(key)
                     .setInitialBalance(new Hbar(1))
                     .execute(testEnv.client);
 
@@ -74,7 +74,7 @@ class TokenFreezeIntegrationTest {
             var key = PrivateKey.generateED25519();
 
             var response = new AccountCreateTransaction()
-                    .setKey(key)
+                    .setKeyWithoutAlias(key)
                     .setInitialBalance(new Hbar(1))
                     .execute(testEnv.client);
 
@@ -137,7 +137,7 @@ class TokenFreezeIntegrationTest {
             var key = PrivateKey.generateED25519();
 
             var response = new AccountCreateTransaction()
-                    .setKey(key)
+                    .setKeyWithoutAlias(key)
                     .setInitialBalance(new Hbar(1))
                     .execute(testEnv.client);